+++ src/main/kotlin/com/asyncapi/plugin/idea/extensions/web/AsyncAPIBrowserUrlProvider.kt
@@ -1,41 +1,27 @@
 package com.asyncapi.plugin.idea.extensions.web
 
-import com.asyncapi.plugin.idea._core.xpath.JsonFileXPath
-import com.asyncapi.plugin.idea._core.xpath.YamlFileXPath
-import com.asyncapi.plugin.idea.extensions.inspection.AsyncAPISchemaDetector
+import com.asyncapi.plugin.idea._core.AsyncAPISpecificationRecognizer
 import com.intellij.ide.browsers.OpenInBrowserRequest
 import com.intellij.ide.browsers.WebBrowserUrlProvider
-import com.intellij.json.psi.JsonFile
+import com.intellij.openapi.components.service
 import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.psi.PsiFile
 import com.intellij.util.Url
-import org.jetbrains.yaml.psi.YAMLFile
 
+/**
+ * @author Pavel Bodiachevskii
+ * @since 1.1.0
+ */
 class AsyncAPIBrowserUrlProvider: WebBrowserUrlProvider() {
 
-    private val asyncApiSchemaDetector = AsyncAPISchemaDetector()
-    private val staticServer = StaticServer()
+    private val asyncAPISpecificationRecognizer = service<AsyncAPISpecificationRecognizer>()
+    private val urlProvider = service<UrlProvider>()
 
     override fun canHandleElement(request: OpenInBrowserRequest): Boolean {
-        if (request.file is JsonFile) {
-            return JsonFileXPath.findText(
-                request.file as JsonFile,
-                "$.asyncapi",
-                false
-            ).isNotEmpty()
-        } else if (request.file is YAMLFile) {
-            return YamlFileXPath.findText(
-                request.file as YAMLFile,
-                "$.asyncapi",
-                false
-            ).isNotEmpty()
-        }
-
-        return false
+        return asyncAPISpecificationRecognizer.isSpecification(request.file)
     }
 
     override fun getUrl(request: OpenInBrowserRequest, file: VirtualFile): Url? {
-        return staticServer.getUrl(request, file)
+        return urlProvider.render(request, file)
     }
 
 }